requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-tables2==2.4.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
outcome==1.2.0
packaging==21.3
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
webdriver-manager==3.8.5
wsproto==1.2.0
